Decoupled logic into different files for clarity and consistency
